Update based on PR comments

+++ db/migrate/20220226172813_add_ndbn_membership_id_to_organizations.rb
@@ -1,6 +1,6 @@
 class AddNdbnMembershipIdToOrganizations < ActiveRecord::Migration[6.1]
   def change
-    add_column :organizations, :ndbn_member_id, :integer
+    add_column :organizations, :ndbn_member_id, :bigint
     add_foreign_key :organizations, :ndbn_members, primary_key: :ndbn_member_id, validate: false
   end
 end
